requirements.txt
groq==1.0.0
openai==2.26.0
pandas==2.3.3
python-dotenv==1.2.2
requests==2.32.5
streamlit==1.55.0
yfinance==1.2.0

